pyproject.toml
[tool.poetry]
name = "tg-login"
version = "0.0.1-pre"
description = "A command line tool to login into Telegram with user or bot accounts."
authors = ["aahnik <daw@aahnik.dev>"]
license = "MIT"
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "^0.17.0"
Telethon = "^1.21.1"
typer = "^0.3.2"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
pytest = "^6.2.3"
tox = "^3.23.0"
pylint = "^2.7.4"
isort = "^5.8.0"

[tool.poetry.scripts]
tg-login = 'tg_login.cli:app'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.isort]
profile = "black"

[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | foo.py           # also separately exclude a file named foo.py in
                     # the root of the project
)
'''